requirements.txt
Django==5.0.6
djangorestframework
djangorestframework-simplejwt
channels
daphne
whitenoise
Pillow
python-telegram-bot==20.0
psycopg2-binary
corsheaders
gunicorn